backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pandas==2.2.3
openpyxl==3.1.5
pdfminer.six==20240706
tenacity==9.0.0
httpx==0.27.2
email-validator==2.3.0
python-multipart==0.0.20
pytest==7.4.3
pytest-asyncio==0.21.1
